refactor: 400 - rename helper file
