+++ src/widgets/courses-school/ui/courses.tsx
@@ -3,7 +3,7 @@ import { CourseCard } from './CourseCard';
 import { MAX_COURSE_COUNT } from '../model/constants';
 import { COURSE_STALE_AFTER_DAYS, ROUTES } from '@/app/const';
 import { courses } from '@/app/services/data';
-import { getActualData } from '@/shared/helpers/getActualDataList';
+import { getActualData } from '@/shared/helpers/getActualData';
 import { useWindowSize } from '@/shared/hooks/use-window-size';
 import { ArrowIcon, RsBanner } from '@/shared/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
