+++ src/widgets/events/ui/events.tsx
@@ -4,7 +4,7 @@ import dayjs from 'dayjs';
 import { events } from '../events.data';
 import { EventCard, EventCardProps } from '@/entities/events';
 import photo3 from '@/shared/assets/photo-3.webp';
-import { getActualData } from '@/shared/helpers/getActualDataList';
+import { getActualData } from '@/shared/helpers/getActualData';
 import Image from '@/shared/ui/image';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { SectionLabel } from '@/shared/ui/section-label';
